requirements.txt
pip
openai
langchain
python-dotenv
beautifulsoup4
requests
tiktoken
sentence-transformers
evaluate
mauve-text
rouge_score
absl-py
loguru
scikit-learn
numpy
pdfminer.six
einops